pyproject.toml
[tool.poetry]
name = "borsh"
version = "0.1.0"
description = "Python implementation of Borsh serialization"
authors = ["kevinheavey <kevinheavey123@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.3"
construct-typing = "^0.5.1"
sumtypes = "^0.1a5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
wemake-python-styleguide = "^0.15.3"
mypy = "^0.910"
black = "^21.7b0"
pydocstyle = "^6.1.1"
isort = "^5.9.3"
coverage = {extras = ["toml"], version = "^5.5"}
pytest-cov = "^2.12.1"
mkdocs = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["borsh"]

[tool.coverage.report]
show_missing = true
fail_under = 100
